Store errors in the appropriate barcode field for easier inspection by consumers.
Do not fail on first extraction/parse error in NewBarcode(). Instead, extract and parse all that can be and store any errors in the appropriate barcode field. Only fail if the barcode data as a whole is invalid.

Add new convenience method `IsDateError()` to allow consumers to easily determine if an error contains a date related error.